Revert "xiph demux: _correctly_ remove warning"
Sorry about that

This reverts commit 020620ba69b19600fe7b06c8fa20cbfe5b50fe49.

+++ modules/demux/xiph.h
@@ -48,10 +48,8 @@ static inline int xiph_SplitHeaders(unsigned packet_size[], void *packet[], unsi
         }
         size += packet_size[i];
     }
-
-    if ((uint8_t)(end - current) < size)
+    if (end - current < (int)size)
         return VLC_EGENERIC;
-
     packet_size[count - 1] = end - current - size;
 
     /* Copy the payloads */
